[Refactor] JSONDecoder 디코딩 비용 과다로 인한 성능 병목 개선 #134
