--- SantaManito-iOS/SantaManito-iOS/Networks/Foundation/Handler/ErrorHandler.swift
@@ -1,33 +1,0 @@
-//
-//  ErrorHandler.swift
-//  SantaManito-iOS
-//
-//  Created by 류희재 on 10/2/24.
-//
-
-import Foundation
-import Combine
-
-//class ErrorHandler {
-//    static let shared = ErrorHandler()
-//    private init () {}
-//    
-//    private let loggingHandler = NetworkLogHandler.shared
-//    
-//    func handleError<T: URLRequestTargetType>(_ target: T, error: SMNetworkError) -> SMNetworkError {
-////        guard let smNetworkError = error as? SMNetworkError else {
-////            return .unknown
-////        }
-//        loggingHandler.responseError(target, result: error)
-//        
-//        return error
-//    }
-//}
-
-//struct ErrorHandler {
-//    static func handleError<T: URLRequestTargetType>(_ target: T, error: SMNetworkError) -> SMNetworkError {
-//        NetworkLogHandler.responseError(target, result: error)
-//        return error
-//    }
-//}
-
